# packages/contracts/openapi/sfs-console.v1.yaml
openapi: 3.1.0
info:
  title: SFS Console API
  version: 0.1.0
paths:
...
  /characters:
    get:
      operationId: listCharacters
      responses:
        "200":
          description: Scanned reusable character registry
    post:
      operationId: createCharacterTemplate
      responses:
        "200":
          description: Created character template files under characters/{slug}
        "422":
          description: Invalid slug, missing fields, or character already exists
...
  /requests/production/preview:
    post:
      operationId: previewProductionRequest
      responses:
        "200":
          description: Paperclip-ready markdown preview
        "422":
          description: Missing required request fields
  /requests/production:
    get:
      operationId: listProductionRequests
      responses:
        "200":
          description: Recent persisted production request drafts
    post:
      operationId: createProductionRequest
      responses:
        "200":
          description: Persisted production request draft with markdown and audit log
        "422":
          description: Missing required request fields
  /requests/production/{request_id}/paperclip:
    post:
      operationId: sendProductionRequestToPaperclip
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Production request linked to a Paperclip issue
        "503":
          description: Paperclip integration is not configured
...